docs: compodoc
