+++ .gitignore
@@ -43,3 +43,5 @@ Thumbs.db
 
 .angular
 .env
+
+documentation
