+++ libs/web/chat/feature-conversation-list/src/lib/conversation-list/is-active.pipe.ts
@@ -1,0 +1,16 @@
+import { inject,Pipe, PipeTransform } from "@angular/core";
+import { ChatFacade, Conversation } from "@chat-app/domain";
+
+@Pipe({
+    standalone: true,
+    name: 'isActive'
+  })
+  export class IsActivePipe implements PipeTransform {
+  
+    private readonly selectedConversation = inject(ChatFacade).selectedConversation;
+  
+    transform(conversation: Conversation): boolean {
+      return conversation.conversationId === this.selectedConversation()?.conversationId;
+    }
+  }
+
