+++ tsconfig.doc.json
@@ -1,0 +1,4 @@
+{
+    "include": ["apps/web/src/**/*.ts"],
+    "exclude": ["**/test.ts", "**/*.spec.ts", "apps/web/src/app.routes.ts"]
+}
